Updates to mar entities
