--- misc/schemas/member-anniversary-roles.jsonc
@@ -1,23 +1,0 @@
-[
-    {
-        "_id": "e47ac83a-a643-461e-8960-344aed6ab391",
-        "+guild": "ObjectId(660711235766976553)",
-        "alias": "60glw3",
-        "year": 1,
-        "role": "867927527245615144"
-    },
-    {
-        "_id": "e47ac83a-a643-461e-8960-344aed6ab391",
-        "+guild": "ObjectId(660711235766976553)",
-        "alias": "70gl23",
-        "year": 2,
-        "role": "883371862418145301"
-    },
-    {
-        "_id": "e47ac83a-a643-461e-8960-344aed6ab391",
-        "+guild": "ObjectId(660711235766976553)",
-        "alias": "90gld3",
-        "year": 3,
-        "role": "867927389236248586"
-    }
-]
